refactor(DailyActivities): 함수, 타입 분리
