+++ app/src/Profile/dashboard-contents/General/DailyActivities/types/DailyActivityScore.d.ts
@@ -1,0 +1,4 @@
+export type DailyActivityScore = {
+  date: Date;
+  score: number;
+};
