+++ app/src/Profile/dashboard-contents/General/DailyActivities/types/DateWithScore.d.ts
@@ -1,0 +1,4 @@
+export type DateWithScore = {
+  date: Date;
+  score: number;
+};
